+++ packages/kit/types/index.d.ts
@@ -1982,7 +1982,11 @@ declare module '@sveltejs/kit' {
 		/** Preflight checks */
 		preflight(schema: StandardSchemaV1<Input, any>): RemoteForm<Input, Output>;
 		/** Validate the form contents programmatically */
-		validate(options?: { includeUntouched?: boolean }): Promise<void>;
+		validate(options?: {
+			includeUntouched?: boolean;
+			/** Perform validation as if the form was submitted by the given button. */
+			submitter?: HTMLButtonElement | HTMLInputElement;
+		}): Promise<void>;
 		/** The result of the form submission */
 		get result(): Output | undefined;
 		/** The number of pending submissions */
@@ -3124,7 +3128,7 @@ declare module '$app/server' {
 	 *
 	 * @since 2.27
 	 */
-	export function form<Output>(fn: () => Output): RemoteForm<void, Output>;
+	export function form<Output>(fn: () => MaybePromise<Output>): RemoteForm<void, Output>;
 	/**
 	 * Creates a form object that can be spread onto a `<form>` element.
 	 *