+++ packages/package/test/fixtures/resolve-alias/expected/index.js
@@ -1,9 +1,0 @@
-export { default as Test } from './Test.svelte';
-export * from "./sub/foo";
-import * as Utils from "./utils/index";
-export { Utils };
-import baz1, { baz } from './baz';
-import baz2 from './baz';
-export { baz, baz1, baz2 };
-// @ts-expect-error
-export { X } from '$libre';